Create a crate, relax solana dependencies in crate (#118)
