+++ token-lending/cli/Cargo.toml
@@ -16,7 +16,7 @@ solana-client = "=1.9.18"
 solana-logger = "=1.9.18"
 solana-sdk = "=1.9.18"
 solana-program = "=1.9.18"
-solend-program = { path="../program", features = [ "no-entrypoint" ] }
+solend-sdk = { path="../sdk" }
 spl-token = { version = "3.2.0", features=["no-entrypoint"] }
 spl-associated-token-account = "1.0.3"
