+++ token-lending/program/src/processor.rs
@@ -13,16 +13,14 @@ use crate::{
         ReserveCollateral, ReserveConfig, ReserveLiquidity,
     },
 };
-use num_traits::FromPrimitive;
 use pyth_sdk_solana::{self, state::ProductAccount};
 use solana_program::{
     account_info::{next_account_info, AccountInfo},
-    decode_error::DecodeError,
     entrypoint::ProgramResult,
     instruction::{get_stack_height, Instruction, TRANSACTION_LEVEL_STACK_HEIGHT},
     msg,
     program::{invoke, invoke_signed},
-    program_error::{PrintProgramError, ProgramError},
+    program_error::ProgramError,
     program_pack::{IsInitialized, Pack},
     pubkey::Pubkey,
     sysvar::instructions::{load_current_index_checked, load_instruction_at_checked},
@@ -32,22 +30,13 @@ use solana_program::{
         Sysvar,
     },
 };
+use solend_sdk::{switchboard_v2_devnet, switchboard_v2_mainnet};
 use spl_token::state::Mint;
 use std::{cmp::min, result::Result};
 use switchboard_program::{
     get_aggregator, get_aggregator_result, AggregatorState, RoundResult, SwitchboardAccountType,
 };
 use switchboard_v2::AggregatorAccountData;
-
-/// Mainnet program id for Switchboard v2.
-pub mod switchboard_v2_mainnet {
-    solana_program::declare_id!("SW1TCH7qEPTdLsDHRgPuMQjbQxKdH2aBStViMFnt64f");
-}
-
-/// Devnet program id for Switchboard v2.
-pub mod switchboard_v2_devnet {
-    solana_program::declare_id!("2TfB33aLaneQb5TNVwyDz3jSZXS6jdW2ARw1Dgf84XCG");
-}
 
 /// Processes an instruction
 pub fn process_instruction(
@@ -2947,13 +2936,4 @@ struct TokenBurnParams<'a: 'b, 'b> {
     authority: AccountInfo<'a>,
     authority_signer_seeds: &'b [&'b [u8]],
     token_program: AccountInfo<'a>,
-}
-
-impl PrintProgramError for LendingError {
-    fn print<E>(&self)
-    where
-        E: 'static + std::error::Error + DecodeError<E> + PrintProgramError + FromPrimitive,
-    {
-        msg!(&self.to_string());
-    }
 }
